Add swift mailer - contact page and confirmation user registration

+++ App/Backend/Modules/Posts/Views/index.php
@@ -22,11 +22,13 @@
         <td>
             <a href="post-update-<?= $post['id'] ?>.html">
                 <i class="fas fa-pen"></i> Modifier
-            </a> 
-            <a href="post-delete-<?= $post['id'] ?>.html" 
-                onclick="return confirm('Voulez-vous supprimer cet article?')">
-                <i class="fas fa-trash ml-4"></i> Supprimer
             </a>
+            <?php if($_SESSION['role_name'] == "Super Admin"): ?> 
+                <a href="post-delete-<?= $post['id'] ?>.html" 
+                    onclick="return confirm('Voulez-vous supprimer cet article?')">
+                    <i class="fas fa-trash ml-4"></i> Supprimer
+                </a>
+            <?php endif; ?>
         </td>
     </tr>
     <?php endforeach; ?>
